Move ApplicationConfiguration to configuration package
