+++ src/main/java/uk/gov/companieshouse/account/validator/service/impl/AccountValidatorImpl.java
@@ -3,7 +3,7 @@ package uk.gov.companieshouse.account.validator.service.impl;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.companieshouse.account.validator.ApplicationConfiguration;
+import uk.gov.companieshouse.account.validator.configuration.ApplicationConfiguration;
 import uk.gov.companieshouse.account.validator.model.validation.ValidationRequest;
 import uk.gov.companieshouse.account.validator.service.AccountValidator;
 import uk.gov.companieshouse.account.validator.service.FelixValidationService;
